requirements.txt
google-api-python-client
google-genai  # Changed from google-generativeai
supabase
python-dotenv
requests
websockets  # Add this